refactor (post): move form category to separate component
